feat: implement QR context for centralized state management and refactor QR forms to use context
