+++ browser/src/components/CustomQr/CustomQr.tsx
@@ -12,160 +12,36 @@ import { Button, ButtonContainer } from "../../styled/styled-button.ts";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faDownload, faGear } from "@fortawesome/free-solid-svg-icons";
 import { faCopy } from "@fortawesome/free-regular-svg-icons";
-import { loadExtensionSettings } from "../../utils/browser-storage.ts";
-import {
-  copyQrCodeToClipboard,
-  downloadQrCodeAsPng,
-  generateSvgContent,
-} from "../../utils/qr-code-gen.ts";
-import { openExtensionSettingsPage } from "../../utils/browser-runtime.ts";
-import { useEffect, useState, useCallback } from "react";
-import {
-  QrDataType,
-  WifiData,
-  ContactData,
-  formatQrContent,
-} from "../../utils/qr-data-formatters.ts";
+import { QrProvider, useQrContext } from "../../contexts/QrContext.tsx";
+import { QrDataType } from "../../utils/qr-data-formatters.ts";
 import QrTypeSelector from "./QrTypes/QrTypeSelector.tsx";
 import TextQrForm from "./QrTypes/TextQrForm.tsx";
 import WifiQrForm from "./QrTypes/WifiQrForm.tsx";
 import ContactQrForm from "./QrTypes/ContactQrForm.tsx";
 
-const CustomQr = () => {
-  const TEXT_PLACEHOLDER = "Set your QR code content here";
-  const WIFI_PLACEHOLDER = {
-    ssid: "Network Name",
-    password: "Password",
-    securityType: "WPA" as const,
-  };
-  const CONTACT_PLACEHOLDER = {
-    name: "John Doe",
-    phone: "+1234567890",
-    email: "john.doe@example.com",
-    organization: "",
-    title: "",
-    website: "",
-  };
-
-  // Initialize the state with a default value
-  const [dataType, setDataType] = useState<QrDataType>(QrDataType.CLEAR_TEXT);
-  const [textContent, setTextContent] = useState(TEXT_PLACEHOLDER);
-  const [wifiData, setWifiData] = useState<WifiData>(WIFI_PLACEHOLDER);
-  const [contactData, setContactData] =
-    useState<ContactData>(CONTACT_PLACEHOLDER);
-  const [qrCodeSvg, setQrCodeSvg] = useState("");
-
-  const getFormattedContent = useCallback(() => {
-    switch (dataType) {
-      case QrDataType.CLEAR_TEXT:
-        return textContent || TEXT_PLACEHOLDER;
-      case QrDataType.WIFI:
-        return formatQrContent(QrDataType.WIFI, wifiData);
-      case QrDataType.CONTACT:
-        return formatQrContent(QrDataType.CONTACT, contactData);
-      default:
-        return textContent;
-    }
-  }, [dataType, textContent, wifiData, contactData]);
-
-  const copyToClipboard = async () => {
-    const extensionSettings = await loadExtensionSettings();
-    const qrContent = getFormattedContent();
-    const svg = await generateSvgContent(
-      qrContent,
-      extensionSettings.qrCodeDownloadSize,
-      extensionSettings.displayLogo,
-    );
-    copyQrCodeToClipboard(svg, extensionSettings.qrCodeDownloadSize).then();
-  };
-
-  const downloadQrCode = async () => {
-    const extensionSettings = await loadExtensionSettings();
-    const qrContent = getFormattedContent();
-    const svg = await generateSvgContent(
-      qrContent,
-      extensionSettings.qrCodeDownloadSize,
-      extensionSettings.displayLogo,
-    );
-    downloadQrCodeAsPng(svg, extensionSettings.qrCodeDownloadSize).then();
-  };
-
-  const openExtensionSettings = () => {
-    openExtensionSettingsPage().then();
-  };
-
-  const updateWifiData = (field: keyof WifiData, value: string) => {
-    setWifiData((prev) => ({
-      ...prev,
-      [field]:
-        field === "securityType" ? (value as "WPA" | "WEP" | "nopass") : value,
-    }));
-  };
-
-  const updateContactData = (field: keyof ContactData, value: string) => {
-    setContactData((prev) => ({
-      ...prev,
-      [field]: value,
-    }));
-  };
-  
-  // Generate QR code SVG when relevant data changes
-  useEffect(() => {
-    let isMounted = true;
-    
-    const generateQrCode = async () => {
-      try {
-        const extensionSettings = await loadExtensionSettings();
-        const qrContent = getFormattedContent();
-        const svg = await generateSvgContent(
-          qrContent,
-          400,
-          extensionSettings.displayLogo
-        );
-        
-        if (isMounted) {
-          setQrCodeSvg(svg);
-        }
-      } catch (error) {
-        console.error("Error generating QR code:", error);
-      }
-    };
-    
-    generateQrCode();
-    
-    return () => {
-      isMounted = false;
-    };
-  }, [getFormattedContent]);
+// Component that uses the QR context
+const QrContent = () => {
+  const {
+    dataType,
+    qrCodeSvg,
+    copyToClipboard,
+    downloadQrCode,
+    openExtensionSettings,
+  } = useQrContext();
 
   const renderQrForm = () => {
     switch (dataType) {
       case QrDataType.CLEAR_TEXT:
-        return (
-          <TextQrForm
-            textContent={textContent}
-            setTextContent={setTextContent}
-            placeholder={TEXT_PLACEHOLDER}
-          />
-        );
-
+        return <TextQrForm />;
       case QrDataType.WIFI:
-        return (
-          <WifiQrForm wifiData={wifiData} updateWifiData={updateWifiData} />
-        );
-
+        return <WifiQrForm />;
       case QrDataType.CONTACT:
-        return (
-          <ContactQrForm
-            contactData={contactData}
-            updateContactData={updateContactData}
-          />
-        );
-
+        return <ContactQrForm />;
       default:
         return null;
     }
   };
+
   return (
     <CustomQrPageContainer>
       <TapToQrHeader title="TapToQR | Create" />
@@ -173,8 +49,7 @@ const CustomQr = () => {
       <CustomQrContainer>
         <FormSideContainer>
           <FormContainer>
-            <QrTypeSelector dataType={dataType} setDataType={setDataType} />
-
+            <QrTypeSelector />
             <FormSection>{renderQrForm()}</FormSection>
           </FormContainer>
         </FormSideContainer>
@@ -209,6 +84,14 @@ const CustomQr = () => {
         </QrSideContainer>
       </CustomQrContainer>
     </CustomQrPageContainer>
+  );
+};
+
+const CustomQr = () => {
+  return (
+    <QrProvider>
+      <QrContent />
+    </QrProvider>
   );
 };
