+++ browser/src/utils/qr-data-formatters.ts
